Change float positions for qt to int

+++ compass.py
@@ -49,7 +49,7 @@ class CompassWidget(QWidget):
         
             if i % 45 == 0:
                 painter.drawLine(0, -40, 0, -50)
-                painter.drawText(-metrics.width(self._pointText[i])/2.0, -52,
+                painter.drawText(int(-metrics.width(self._pointText[i])/2.0), -52,
                                  self._pointText[i])
             else:
                 painter.drawLine(0, -45, 0, -50)
